app/requirements.txt
boto3==1.36.14
cfl-common
codeforlife-portal
django-redis==5.4.0
gunicorn==23.0.0
mypy-boto3-s3==1.36.9
python-dotenv==1.0.1
rapid-router
redis==5.2.0
requests-toolbelt==1.0.0
uvicorn-worker==0.2.0
